# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API документация",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/v1/tenders": {
            "get": {
                "tags": [
                    "Тендеры"
                ],
                "summary": "Получение списка тендеров",
                "operationId": "4157751033dde5642f56dc97b4007200",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Фильтр по названию тендера",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Тендер"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Фильтр по дате (формат Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2022-08-14"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записей на страницу",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "text",
                                                        "example": "Тендера на стройматериалы"
                                                    },
                                                    "status": {
                                                        "type": "text",
                                                        "example": "Открыто"
                                                    },
                                                    "number": {
                                                        "type": "text",
                                                        "example": "1724-2"
                                                    },
                                                    "external_code": {
                                                        "type": "integer",
                                                        "example": "53228"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-06-01T18:36:50.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "description": "Навигационные ссылки для пагинации",
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/tenders?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/tenders?page=1"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "description": "Метаинформация о пагинации",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "links": {
                                                    "description": "Массив ссылок на страницы",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": "http://127.0.0.1:8000/api/v1/tenders?page=1",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "1"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/v1/tenders"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Тендеры"
                ],
                "summary": "Создание тендера",
                "operationId": "841b5301255554aa2b4fde513cff04da",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "text",
                                                "example": "Тендера на стройматериалы"
                                            },
                                            "status": {
                                                "type": "text",
                                                "example": "Открыто"
                                            },
                                            "number": {
                                                "type": "text",
                                                "example": "1724-2"
                                            },
                                            "external_code": {
                                                "type": "integer",
                                                "example": "53228"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "text",
                                                    "example": "Тендера на стройматериалы"
                                                },
                                                "status": {
                                                    "type": "text",
                                                    "example": "Открыто"
                                                },
                                                "number": {
                                                    "type": "text",
                                                    "example": "1724-2"
                                                },
                                                "external_code": {
                                                    "type": "integer",
                                                    "example": "53228"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-06-01T18:36:50.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/tenders/{id}": {
            "get": {
                "tags": [
                    "Тендеры"
                ],
                "summary": "Получение тендера",
                "operationId": "060137e9e759e8b22c03950e225cb544",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID тендера",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "text",
                                                    "example": "Тендера на стройматериалы"
                                                },
                                                "status": {
                                                    "type": "text",
                                                    "example": "Открыто"
                                                },
                                                "number": {
                                                    "type": "text",
                                                    "example": "1724-2"
                                                },
                                                "external_code": {
                                                    "type": "integer",
                                                    "example": "53228"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-06-01T18:36:50.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Аутентификация через токен Laravel Sanctum",
                "bearerFormat": "Laravel Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}